refactor mcp sandbox read helpers
